refactor/#2 - 쿼리 최적화 적용

+++ src/main/java/lucky/seven/kanbanbuzz/service/ColumnService.java
@@ -58,8 +58,9 @@ public class ColumnService {
 
         checkUser(user);
 
-        columnRepository.findByIdAndBoardId(columnId, boardId).orElseThrow(() ->
-                new UserException(ErrorType.INVALID_COLUMN));
+        if(!columnRepository.existsByIdAndBoardId(columnId, boardId)) {
+            throw new UserException(ErrorType.INVALID_COLUMN);
+        }
 
         columnRepository.deleteById(columnId);
     }
